Move WebhooksByEventTypeLoader inside dataloaders module
