+++ saleor/graphql/webhook/dataloaders/__init__.py
@@ -2,6 +2,7 @@ from .models import (
     PayloadByIdLoader,
     WebhookEventsByWebhookIdLoader,
     WebhooksByAppIdLoader,
+    WebhooksByEventTypeLoader,
 )
 from .pregenerated_payload_for_checkout_tax import (
     PregeneratedCheckoutTaxPayloadsByCheckoutTokenLoader,
@@ -9,7 +10,7 @@ from .pregenerated_payload_for_checkout_tax import (
 from .pregenerated_payloads_for_checkout_filter_shipping_methods import (
     PregeneratedCheckoutFilterShippingMethodPayloadsByCheckoutTokenLoader,
 )
-from .utils import PayloadsRequestContextByEventTypeLoader, WebhooksByEventTypeLoader
+from .utils import PayloadsRequestContextByEventTypeLoader
 
 __all__ = [
     "PayloadByIdLoader",
