+++ saleor/graphql/webhook/dataloaders/models.py
@@ -1,7 +1,12 @@
 from collections import defaultdict
 
 from ....core.models import EventPayload
+from ....webhook.event_types import WebhookEventAsyncType
 from ....webhook.models import Webhook, WebhookEvent
+from ....webhook.utils import (
+    calculate_webhooks_for_multiple_events,
+)
+from ...app.dataloaders import ActiveAppByIdLoader
 from ...core.dataloaders import DataLoader
 
 
@@ -70,3 +75,68 @@ class ActiveWebhooksByIdLoader(DataLoader):
             .in_bulk(keys)
         )
         return [webhooks_map.get(webhook_id) for webhook_id in keys]
+
+
+class WebhooksByEventTypeLoader(DataLoader):
+    context_key = "webhooks_by_event_type"
+
+    def batch_load(self, keys):
+        set_event_types = set(keys)
+        if set_event_types.intersection(WebhookEventAsyncType.ALL):
+            set_event_types.add(WebhookEventAsyncType.ANY)
+
+        def fetch_webhooks(
+            webhook_events_by_event_types, set_event_types=set_event_types
+        ):
+            events_types_by_webhook_id_map: dict[int, set[str]] = defaultdict(set)
+            webhooks_ids = set()
+            for webhook_events_by_event_type in webhook_events_by_event_types:
+                for webhook_event in webhook_events_by_event_type:
+                    webhooks_id = webhook_event.webhook_id
+                    events_types_by_webhook_id_map[webhook_event.webhook_id].add(
+                        webhook_event.event_type
+                    )
+                    webhooks_ids.add(webhooks_id)
+
+            def fetch_apps(
+                webhooks,
+                set_event_types=set_event_types,
+                events_types_by_webhook_id_map=events_types_by_webhook_id_map,
+            ):
+                app_ids = {webhook.app_id for webhook in webhooks}
+
+                def return_webhooks_by_event_type(
+                    apps,
+                    set_event_types=set_event_types,
+                    events_types_by_webhook_id_map=events_types_by_webhook_id_map,
+                    webhooks=webhooks,
+                ):
+                    apps_by_id = {app.id: app for app in apps}
+                    webhooks_by_event_type_map = calculate_webhooks_for_multiple_events(
+                        set_event_types,
+                        apps_by_id,
+                        webhooks,
+                        events_types_by_webhook_id_map,
+                    )
+                    return [
+                        webhooks_by_event_type_map.get(event_type, [])
+                        for event_type in keys
+                    ]
+
+                return (
+                    ActiveAppByIdLoader(self.context)
+                    .load_many(app_ids)
+                    .then(return_webhooks_by_event_type)
+                )
+
+            return (
+                ActiveWebhooksByIdLoader(self.context)
+                .load_many(webhooks_ids)
+                .then(fetch_apps)
+            )
+
+        return (
+            WebhookEventsByEventTypeLoader(self.context)
+            .load_many(set_event_types)
+            .then(fetch_webhooks)
+        )
