+++ saleor/graphql/webhook/dataloaders/pregenerated_payload_for_checkout_tax.py
@@ -19,10 +19,10 @@ from ...checkout.dataloaders.checkout_infos import (
 from ...core.dataloaders import DataLoader
 from ..subscription_payload import generate_payload_promise_from_subscription
 from ..utils import get_subscription_query_hash
+from .models import WebhooksByEventTypeLoader
 from .utils import (
     AppsByEventTypeLoader,
     PayloadsRequestContextByEventTypeLoader,
-    WebhooksByEventTypeLoader,
 )
