+++ saleor/graphql/webhook/dataloaders/utils.py
@@ -1,16 +1,12 @@
 from collections import defaultdict
 from typing import Optional
 
-from ....webhook.event_types import WebhookEventAsyncType
-from ....webhook.utils import (
-    calculate_webhooks_for_multiple_events,
-)
-from ...app.dataloaders import ActiveAppByIdLoader, AppByIdLoader
+from ...app.dataloaders import AppByIdLoader
 from ...core import SaleorContext
 from ...core.dataloaders import DataLoader
 from ...utils import get_user_or_app_from_context
 from ..subscription_payload import initialize_request
-from .models import ActiveWebhooksByIdLoader, WebhookEventsByEventTypeLoader
+from .models import WebhooksByEventTypeLoader
 
 
 class PayloadsRequestContextByEventTypeLoader(DataLoader):
@@ -62,69 +58,4 @@ class AppsByEventTypeLoader(DataLoader):
             WebhooksByEventTypeLoader(self.context)
             .load_many(keys)
             .then(return_apps_for_webhooks)
-        )
-
-
-class WebhooksByEventTypeLoader(DataLoader):
-    context_key = "webhooks_by_event_type"
-
-    def batch_load(self, keys):
-        set_event_types = set(keys)
-        if set_event_types.intersection(WebhookEventAsyncType.ALL):
-            set_event_types.add(WebhookEventAsyncType.ANY)
-
-        def fetch_webhooks(
-            webhook_events_by_event_types, set_event_types=set_event_types
-        ):
-            events_types_by_webhook_id_map: dict[int, set[str]] = defaultdict(set)
-            webhooks_ids = set()
-            for webhook_events_by_event_type in webhook_events_by_event_types:
-                for webhook_event in webhook_events_by_event_type:
-                    webhooks_id = webhook_event.webhook_id
-                    events_types_by_webhook_id_map[webhook_event.webhook_id].add(
-                        webhook_event.event_type
-                    )
-                    webhooks_ids.add(webhooks_id)
-
-            def fetch_apps(
-                webhooks,
-                set_event_types=set_event_types,
-                events_types_by_webhook_id_map=events_types_by_webhook_id_map,
-            ):
-                app_ids = {webhook.app_id for webhook in webhooks}
-
-                def return_webhooks_by_event_type(
-                    apps,
-                    set_event_types=set_event_types,
-                    events_types_by_webhook_id_map=events_types_by_webhook_id_map,
-                    webhooks=webhooks,
-                ):
-                    apps_by_id = {app.id: app for app in apps}
-                    webhooks_by_event_type_map = calculate_webhooks_for_multiple_events(
-                        set_event_types,
-                        apps_by_id,
-                        webhooks,
-                        events_types_by_webhook_id_map,
-                    )
-                    return [
-                        webhooks_by_event_type_map.get(event_type, [])
-                        for event_type in keys
-                    ]
-
-                return (
-                    ActiveAppByIdLoader(self.context)
-                    .load_many(app_ids)
-                    .then(return_webhooks_by_event_type)
-                )
-
-            return (
-                ActiveWebhooksByIdLoader(self.context)
-                .load_many(webhooks_ids)
-                .then(fetch_apps)
-            )
-
-        return (
-            WebhookEventsByEventTypeLoader(self.context)
-            .load_many(set_event_types)
-            .then(fetch_webhooks)
         )
